backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-decouple
